Initial Files for Journey to Pluto
If anyone is good at exceptions please help with that. The exception is used in the function at the bottom of the resources class.
This is the very first, uncompleted form of the base classes for our project. No GUI work has yet been done. 
The resource requirements and modifiers may need to be changed so the game can be played in a reasonable amount of time.
No work on multiplayer has yet been done
No testing has yet been done