Delete InsufficientResourcesException
This was previously used in the Resources class in the method useResourcesForTravel. With the addition of methods to Player for use in traveling, safeguards will be in place such that this exception is unnecessary and useResourcesForTravel will no longer check whether there are enough resources. Instead, only planets the player can travel to will be allowed as choices. This means less error messages and is more user-friendly.