Updated to no longer need GUI
This class has been updated to no longer need GUI in its functions:
takeTurn function has been removed. It's functionality requires the GUI and was closer
to the previous project. 

The collectResources() fuction has been broken up into three functions
getMultipliers() which returns an int array of multipliers (for GUI display)
getPossibleResources() which performs the calculations with the multipliers and returns an int array of possible resources
collectResources() which adds resources of a given type to a user's stores 

commenceTravel has likewise been broken up. tryTravel gets the successRate, then the GUI asks if the user wishes to proceed, then
the commenceTravel function is called to attempt travel and deduct resources where successful 

Additionally, the name reference field has been added, put into the constructor, and a getName() function implemented.
Constructor updated to use final ints for number of sides and number of Dice since these values do not change.